requirements.txt
# Core AI/LLM Framework
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
openai>=1.0.0

# API Integrations
googlemaps>=4.0.0
requests>=2.25.0
python-dotenv>=0.19.0

# Data Processing
pandas>=1.5.0
numpy>=1.20.0

# PDF Generation
reportlab>=3.6.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
flask>=2.3.0
pydantic>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.0
geopy>=2.3.0

# Development
pytest>=7.0.0 